refactor: move document verification module
