+++ main.js
@@ -2,6 +2,7 @@ import { initTheme } from './theme.js';
 import { initCalendar, resetAppointmentForm } from './calendar.js';
 import { initClientPortal } from './portal.js';
 
+import { initDocumentVerification } from './src/document-verification.js';
 initTheme();
 
         // Mobile menu toggle
@@ -260,314 +261,4 @@ initTheme();
             initDocumentVerification();
         });
         
-        
-        // Document Pre-Verification Tool functionality
-        function initDocumentVerification() {
-            const dropZone = document.getElementById('drop-zone');
-            const fileInput = document.getElementById('file-input');
-            const browseFilesBtn = document.getElementById('browse-files');
-            const uploadSection = document.getElementById('upload-section');
-            const verificationProgress = document.getElementById('verification-progress');
-            const verificationResults = document.getElementById('verification-results');
-            const progressBar = document.getElementById('progress-bar');
-            const progressPercentage = document.getElementById('progress-percentage');
-            const resultIcon = document.getElementById('result-icon');
-            const resultStatus = document.getElementById('result-status');
-            const documentType = document.getElementById('document-type');
-            const requirementsList = document.getElementById('requirements-list');
-            const recommendations = document.getElementById('recommendations');
-            const verifyAnotherBtn = document.getElementById('verify-another');
-            
-            // If any element doesn't exist, return early
-            if (!dropZone || !fileInput || !browseFilesBtn) {
-                return;
-            }
-            
-            // Make sure all elements are properly styled
-            if (verificationProgress) {
-                verificationProgress.classList.add('hidden');
-            }
-
-            if (verificationResults) {
-                verificationResults.classList.add('hidden');
-            }
-
-            if (uploadSection) {
-                uploadSection.classList.remove('hidden');
-            }
-            
-            // Show the verification tool when available
-            const documentVerificationTool = document.getElementById('document-verification-tool');
-            if (documentVerificationTool) {
-                documentVerificationTool.classList.remove('hidden');
-            }
-            
-            // Ensure drop zone classes are applied
-            if (dropZone) {
-                dropZone.classList.add(
-                    'border-2',
-                    'border-dashed',
-                    'border-mediumgray',
-                    'rounded-sm',
-                    'p-8',
-                    'text-center',
-                    'cursor-pointer',
-                    'transition-colors',
-                    'duration-300',
-                    'bg-white',
-                    'min-h-[250px]',
-                    'box-border',
-                    'w-full'
-                );
-            }
-            
-            // Make sure the file input is hidden
-            if (fileInput) {
-                fileInput.classList.add('hidden');
-            }
-            
-            // Click browse files button
-            if (browseFilesBtn) {
-                browseFilesBtn.addEventListener('click', (e) => {
-                    e.preventDefault();
-                    if (fileInput) fileInput.click();
-                });
-            }
-            
-            // File input change
-            if (fileInput) {
-                fileInput.addEventListener('change', () => {
-                    if (fileInput.files.length > 0) {
-                        handleFile(fileInput.files[0]);
-                    }
-                });
-            }
-            
-            // Drag and drop functionality
-            if (dropZone) {
-                dropZone.addEventListener('dragover', (e) => {
-                    e.preventDefault();
-                    dropZone.classList.add('border-charcoal', 'bg-lightgray', 'dark:bg-darkgray');
-                });
-
-                dropZone.addEventListener('dragleave', () => {
-                    dropZone.classList.remove('border-charcoal', 'bg-lightgray', 'dark:bg-darkgray');
-                });
-
-                dropZone.addEventListener('drop', (e) => {
-                    e.preventDefault();
-                    dropZone.classList.remove('border-charcoal', 'bg-lightgray', 'dark:bg-darkgray');
-
-                    if (e.dataTransfer.files.length > 0) {
-                        handleFile(e.dataTransfer.files[0]);
-                    }
-                });
-                
-                // Add keyboard support for drop zone
-                dropZone.addEventListener('keydown', (e) => {
-                    if (e.key === 'Enter' || e.key === ' ') {
-                        e.preventDefault();
-                        if (fileInput) fileInput.click();
-                    }
-                });
-                
-                // Add click support for drop zone
-                dropZone.addEventListener('click', () => {
-                    if (fileInput) fileInput.click();
-                });
-            }
-            
-            // Verify another document button
-            if (verifyAnotherBtn) {
-                verifyAnotherBtn.addEventListener('click', () => {
-                    if (verificationResults) {
-                        verificationResults.classList.add('hidden');
-                    }
-
-                    if (uploadSection) {
-                        uploadSection.classList.remove('hidden');
-                    }
-                    
-                    if (fileInput) {
-                        fileInput.value = '';
-                    }
-                });
-            }
-            
-            // Handle file upload and verification
-            function handleFile(file) {
-                // Check file type
-                const validTypes = ['application/pdf', 'application/msword', 'application/vnd.openxmlformats-officedocument.wordprocessingml.document', 'image/jpeg', 'image/png'];
-                if (!validTypes.includes(file.type)) {
-                    const uploadError = document.getElementById('upload-error');
-                    if (uploadError) {
-                        uploadError.textContent = 'Please upload a valid document (PDF, DOC, DOCX, JPG, or PNG).';
-                        uploadError.classList.remove('hidden');
-                        uploadError.focus();
-                    }
-                    return;
-                }
-                
-                // Show progress
-                if (uploadSection) {
-                    uploadSection.classList.add('hidden');
-                }
-
-                if (verificationProgress) {
-                    verificationProgress.classList.remove('hidden');
-                }
-                
-                // Simulate progress (in a real app, this would be actual file upload and analysis)
-                let progress = 0;
-                const interval = setInterval(() => {
-                    progress += 5;
-                    
-                    if (progressBar) {
-                        progressBar.style.width = `${progress}%`;
-                    }
-                    
-                    if (progressPercentage) {
-                        progressPercentage.textContent = `${progress}%`;
-                    }
-                    
-                    if (progress >= 100) {
-                        clearInterval(interval);
-                        setTimeout(() => {
-                            if (verificationProgress) {
-                                verificationProgress.classList.add('hidden');
-                            }
-
-                            if (verificationResults) {
-                                verificationResults.classList.remove('hidden');
-                            }
-                            
-                            analyzeDocument(file);
-                        }, 500);
-                    }
-                }, 100);
-                
-            }
-            
-            // Basic client-side document analysis
-            function analyzeDocument(file) {
-                
-                const fileName = file.name.toLowerCase();
-                let documentCategory;
-                let isValid = true;
-                const issues = [];
-                
-                // Announce to screen readers that analysis is complete
-                const resultStatusContainer = document.getElementById('result-status-container');
-                resultStatusContainer.setAttribute('aria-busy', 'false');
-                
-                // Determine document type based on filename
-                if (fileName.includes('power') || fileName.includes('attorney')) {
-                    documentCategory = 'Power of Attorney';
-                } else if (fileName.includes('deed') || fileName.includes('title') || fileName.includes('property')) {
-                    documentCategory = 'Real Estate Document';
-                } else if (fileName.includes('will') || fileName.includes('testament')) {
-                    documentCategory = 'Last Will and Testament';
-                } else if (fileName.includes('affidavit')) {
-                    documentCategory = 'Affidavit';
-                } else {
-                    documentCategory = 'General Document';
-                }
-                
-                if (file.size > 5 * 1024 * 1024) {
-                    isValid = false;
-                    issues.push('File exceeds 5MB size limit');
-                }
-
-                if (documentCategory === 'General Document') {
-                    isValid = false;
-                    issues.push('Unrecognized document type; please include a notarial certificate and signature line');
-                }
-                
-                // Display results
-                documentType.textContent = documentCategory;
-                
-                // Set requirements based on document type
-                requirementsList.innerHTML = '';
-                let reqs = [];
-                
-                switch (documentCategory) {
-                    case 'Power of Attorney':
-                        reqs = [
-                            'Valid government-issued photo ID',
-                            'Original document with signature line',
-                            'All signers must appear in person',
-                            'Witness(es) may be required depending on state law'
-                        ];
-                        break;
-                    case 'Real Estate Document':
-                        reqs = [
-                            'Valid government-issued photo ID for all signers',
-                            'Original document with all pages',
-                            'All signers must appear in person',
-                            'Proper notarial certificate or acknowledgment'
-                        ];
-                        break;
-                    case 'Last Will and Testament':
-                        reqs = [
-                            'Valid government-issued photo ID',
-                            'Original document with signature line',
-                            'Witnesses (typically 2) must be present',
-                            'Self-proving affidavit recommended'
-                        ];
-                        break;
-                    default:
-                        reqs = [
-                            'Valid government-issued photo ID',
-                            'Original document with signature line',
-                            'All signers must appear in person'
-                        ];
-                }
-                
-                reqs.forEach(req => {
-                    const li = document.createElement('li');
-                    li.textContent = req;
-                    requirementsList.appendChild(li);
-                });
-                
-                // Set recommendations
-                if (isValid) {
-                    resultStatusContainer.className = 'p-4 rounded-sm mb-4 bg-green-100';
-                    resultIcon.innerHTML = '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M9 12l2 2 4-4m6 2a9 9 0 11-18 0 9 9 0 0118 0z" />';
-                    resultIcon.classList.add('text-green-600');
-                    resultStatus.textContent = 'Document Ready for Notarization';
-                    resultStatus.classList.add('text-green-800');
-                    
-                    recommendations.innerHTML = `
-                        <p class="mb-2">Your document appears to be ready for notarization. Please bring the following to your appointment:</p>
-                        <ul class="list-disc pl-5 space-y-1">
-                            <li>Original document (all pages)</li>
-                            <li>Valid government-issued photo ID</li>
-                            <li>Any witnesses required for your document type</li>
-                        </ul>
-                    `;
-                } else {
-                    resultStatusContainer.className = 'p-4 rounded-sm mb-4 bg-yellow-100';
-                    resultIcon.innerHTML = '<path stroke-linecap="round" stroke-linejoin="round" stroke-width="2" d="M12 9v2m0 4h.01m-6.938 4h13.856c1.54 0 2.502-1.667 1.732-3L13.732 4c-.77-1.333-2.694-1.333-3.464 0L3.34 16c-.77 1.333.192 3 1.732 3z" />';
-                    resultIcon.classList.add('text-yellow-600');
-                    resultStatus.textContent = 'Document Needs Attention';
-                    resultStatus.classList.add('text-yellow-800');
-                    
-                    let issuesHtml = '';
-                    issues.forEach(issue => {
-                        issuesHtml += `<li>${issue}</li>`;
-                    });
-                    
-                    recommendations.innerHTML = `
-                        <p class="mb-2">Your document needs some adjustments before notarization:</p>
-                        <ul class="list-disc pl-5 space-y-1 text-yellow-800">
-                            ${issuesHtml}
-                        </ul>
-                        <p class="mt-2">We recommend consulting with our notary professionals to ensure your document meets all requirements.</p>
-                    `;
-                }
-                
-            }
-        }
-        
-
